Move ReflectionUtil into util sub-package
